Update install-mxnet-osx-python.sh (#8825)
* Update install-mxnet-osx-python.sh

Fixed handling of MXNET_TAG. If not set latest from master will be cloned to local directory. If a tag is specified, that tag will be used to clone the repository.

Tested with 
MXNET_TAG="1.0.0.rc0"

* Update install-mxnet-osx-python.sh

added following modifications:
a) added selection for reporting list of available tags
b) moved the tag selection into the introduction section

open: still need to resolve proper installation of the build version

* Update install-mxnet-osx-python.sh

Added check if mxnet was already installed on the system (e.g. pip install), as this might interfere with the attempt to download and install a new version.
Performed manual tests, all passing. This is a final commit.

* Update install-mxnet-osx-python.sh

empty change to trigger build.

* Update install-mxnet-osx-python.sh

another dummy change to fight the CI system.