Make MXNetError in Scala Not Private (#11491)
`MXNetError` is thrown in the event of a CUDA error (e.g., `Check failed: err == cudaSuccess ...`). While ideally we'd never encounter this, it happens sometimes, and we'd like to program our scala code to take the appropriate action (i.e., restart the process).

This PR proposes making the `MXNetError` class public, so that we can `catch` it in our calling code.